% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_storemapping_templates.r
\name{create_storemapping_templates}
\alias{create_storemapping_templates}
\title{Create storemapping templates}
\usage{
create_storemapping_templates()
}
\value{
return 2 CSV files stored in working directory, "clientstore.csv" and "storemaster.csv"
}
\description{
create standard data templates for storemapping function
}
\examples{
#suggest to use it under your working directory
setwd("set your working directory here")
create_storemapping_templates()
}
