% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storemapping_core_v1.2.r
\name{storemapping_core}
\alias{storemapping_core}
\title{Nielsen StoreMapping Tool}
\usage{
storemapping_core(Storemaster_Channel_Selection, road_mapping_filter_rank,
  GPS_mapping_filter_rank, GPS_mapping_maxdis)
}
\arguments{
\item{Storemaster_Channel_Selection}{choose channels to match, separate with "|", eg baby|cos|csv|hyper|mini|super|'}

\item{road_mapping_filter_rank}{in store similarity process, for each client store, filter the TopN(road_mapping_filter_rank) closest nielsen stores}

\item{GPS_mapping_filter_rank}{in GPS distance process, for each client store, filter the TopN(GPS_mapping_filter_rank) closest nielsen stores}

\item{GPS_mapping_maxdis}{in GPS distance process, filter the pairs distance should be less than GPS_mapping_maxdis,default = 0.001}
}
\value{
The result will be CSV file named "map_result_final.csv" restored in your working directory
}
\description{
Map LINX stores with RMS stores
}
\details{
must contain type(aka store type) and cityname, as this tool mapping under the hood of these 2 fields
}
\examples{
setwd("setting your working directory that contains with data in accordance with templates")
Storemaster_Channel_Selection = "baby|cos|csv|hyper|mini|super"
road_mapping_filter_rank = 3
GPS_mapping_filter_rank = 2
GPS_mapping_maxdis = 0.001
distype = 'km' #'coord'
map_result_final = storemapping_core(Storemaster_Channel_Selection,
                                     road_mapping_filter_rank,
                                     GPS_mapping_filter_rank,
                                     GPS_mapping_maxdis)
}
